+++ stac_fastapi/api/tests/test_app.py
@@ -17,7 +17,7 @@ from stac_fastapi.api.models import (
     create_get_request_model,
     create_post_request_model,
 )
-from stac_fastapi.extensions.core import (
+from stac_fastapi.extensions import (
     FieldsExtension,
     FilterExtension,
     SearchFilterExtension,