+++ stac_fastapi/extensions/stac_fastapi/extensions/__init__.py
@@ -1,1 +1,41 @@
-"""Extensions submodule."""
+"""stac_api.extensions module."""
+
+from .aggregation import AggregationExtension
+from .bulk_transactions.bulk_transactions import BulkTransactionExtension
+from .collection_search import CollectionSearchExtension, CollectionSearchPostExtension
+from .fields import FieldsExtension
+from .filter import (
+    CollectionSearchFilterExtension,
+    FilterExtension,
+    ItemCollectionFilterExtension,
+    SearchFilterExtension,
+)
+from .free_text import FreeTextAdvancedExtension, FreeTextExtension
+from .pagination import (
+    OffsetPaginationExtension,
+    PaginationExtension,
+    TokenPaginationExtension,
+)
+from .query import QueryExtension
+from .sort import SortExtension
+from .transaction import TransactionExtension
+
+__all__ = (
+    "AggregationExtension",
+    "BulkTransactionExtension",
+    "FieldsExtension",
+    "FilterExtension",
+    "FreeTextExtension",
+    "FreeTextAdvancedExtension",
+    "OffsetPaginationExtension",
+    "PaginationExtension",
+    "QueryExtension",
+    "SortExtension",
+    "TokenPaginationExtension",
+    "TransactionExtension",
+    "CollectionSearchExtension",
+    "CollectionSearchPostExtension",
+    "SearchFilterExtension",
+    "ItemCollectionFilterExtension",
+    "CollectionSearchFilterExtension",
+)