+++ stac_fastapi/extensions/tests/test_filter.py
@@ -6,8 +6,8 @@ from starlette.testclient import TestClient
 
 from stac_fastapi.api.app import StacApi
 from stac_fastapi.api.models import create_get_request_model, create_post_request_model
-from stac_fastapi.extensions.core import FilterExtension
-from stac_fastapi.extensions.core.filter import (
+from stac_fastapi.extensions import FilterExtension
+from stac_fastapi.extensions.filter import (
     CollectionSearchFilterExtension,
     ItemCollectionFilterExtension,
     SearchFilterExtension,