+++ CHANGES.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Changed
+
+- removed the third party and core extensions folders and moved all of the extensions to root. Ensured backwards compatibility and added a deprecation warning. ([#899](https://github.com/stac-utils/stac-fastapi/pull/899))
+
 ### Fixed
 
 - fix mypy type errors in transaction extension for Python 3.14 compatibility (mypy 1.20.0) ([#895](https://github.com/stac-utils/stac-fastapi/pull/895))